Updating Tetris.py
Used to not work, there was a misspelling in rotation in shape.shape % shape.rotation. And then draw grid in line 290 missed the conditions row and column so I removed row and column because I wasn't using them, because [i][j] was already the logic in rows and columns. Game works now, just need to fix the highscore, also edited the score to make it x 100. High score adds on to the previous and is only single digits.